AWSMTestCase - Change to use SetUpClass.
No need to call this with every test case run. Also add option to skip
clearing of output folders.

+++ tests/awsm_test_case.py
@@ -15,7 +15,7 @@ class AWSMTestCase(unittest.TestCase):
     """
 
     @classmethod
-    def setUp(cls):
+    def setUpClass(cls):
         cls.test_dir = os.path.abspath(os.path.join(
             os.path.dirname(awsm.__file__), '..', 'tests'
         ))
@@ -30,15 +30,18 @@ class AWSMTestCase(unittest.TestCase):
         """
         Clean up the output directory
         """
-        folder = os.path.join(self.test_dir, 'RME', 'output')
-        no_delete = os.path.join(folder, '.keep')
-        for the_file in os.listdir(folder):
-            file_path = os.path.join(folder, the_file)
-            if file_path != no_delete:
-                try:
-                    if os.path.isfile(file_path):
-                        os.unlink(file_path)
-                    elif os.path.isdir(file_path):
-                        shutil.rmtree(file_path)
-                except Exception as e:
-                    print(e)
+        if hasattr(self, 'cache_run') and self.cache_run:
+            return
+        else:
+            folder = os.path.join(self.test_dir, 'RME', 'output')
+            no_delete = os.path.join(folder, '.keep')
+            for the_file in os.listdir(folder):
+                file_path = os.path.join(folder, the_file)
+                if file_path != no_delete:
+                    try:
+                        if os.path.isfile(file_path):
+                            os.unlink(file_path)
+                        elif os.path.isdir(file_path):
+                            shutil.rmtree(file_path)
+                    except Exception as e:
+                        print(e)
